+++ appObjects/FlatCAMCNCJob.py
@@ -484,6 +484,8 @@ class CNCJobObject(FlatCAMObj, CNCjob):
         # Include CNC Job Snippets changed
         self.ui.snippets_cb.toggled.connect(self.on_update_source_file)
 
+        self.ui.autolevel_button.clicked.connect(lambda: self.app.levelling_tool.run(toggle=True))
+
         # ###################################### END Signal connections ###############################################
         # #############################################################################################################